% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalizer.R
\name{equalize_IVT}
\alias{equalize_IVT}
\title{Run equalizer on IVT-format arrays.}
\usage{
equalize_IVT(fn_python, dir_bedtools_bin, package_name, fn_vcf_list, affy_bed,
  annot_version, annot_revision, genome_version_UCSC, genome_version_NCBI, chip,
  chip_version, species, organism, author, email, fn_CDF, fn_probe_tab, fn_CEL,
  fn_probeset_csv, dir_out)
}
\arguments{
\item{fn_python}{Path to python executable}

\item{dir_bedtools_bin}{Path to bin folder of bedtools}

\item{package_name}{Name of the new package to create}

\item{fn_vcf_list}{Vector of paths to VCF files}

\item{affy_bed}{path to Affymetrix probe locations in BED format (sourced from Netaffx)}

\item{annot_version}{Affymetrix annotation version for the source chip}

\item{annot_revision}{Affymetrix chip library revision number for the source chip}

\item{genome_version_UCSC}{UCSC Genome version for the source chip}

\item{genome_version_NCBI}{NCI Genome version for the source chip}

\item{chip}{Affymetrix chip identifier for the source chip}

\item{chip_version}{Affymetrix chip version string for the source chip}

\item{species}{Scientific species name for the source chip}

\item{organism}{Common species name for the source chip}

\item{author}{Author name for new package}

\item{email}{Author email for new package}

\item{fn_CDF}{Path to Affymetrix CDF file (sourced from Netaffx)}

\item{fn_probe_tab}{Path to Affymetrix probe_tab file (sourced from Netaffx)}

\item{fn_CEL}{Path to an Affymetrix CEL file of the type produced by the source chip}

\item{fn_probeset_csv}{Path to affymetrix probeset.csv (sourced from Netaffx)}

\item{dir_out}{Folder where new package files will be created}
}
\description{
\code{equalize_IVT} uses one or more VCF files defining the locations of 
  Single Nucleotide Polymorphisms (SNPs) to rewrite Affymetrix array description files 
  for a source package so they exclude any probes which overlap SNPs. 
  The \code{equalize_IVT} function is a wrapper around a Python script.
  
  To run this function, \strong{you must have Python and bedtools installed locally}.
  
  The files where (sourced from Netaffx) is listed are available 
  from Affymetrix's Netaffx website. You may also be able to find copies at 
  davidquigley.com .
}
\seealso{
\code{\link{equalize_gene}} for ST-format arrays (e.g. gene ST 1.1)
}

