% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/equalizer.R
\docType{package}
\name{equalizer}
\alias{check_file}
\alias{equalizer}
\alias{equalizer-package}
\title{equalizer: A package for removing SNP effects from Affymetrix microarrays}
\usage{
check_file(fn)
}
\description{
equalizer: A package for removing SNP effects from Affymetrix microarrays
}

